Organsised java files with subfolders
