+++ QuiRky/app/src/main/AndroidManifest.xml
@@ -22,19 +22,19 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.QuiRky">
         <activity
-            android:name=".LeaderBoardActivity"
+            android:name=".activities.LeaderBoardActivity"
             android:exported="false" />
 	<activity
-            android:name=".delete_Players"
+            android:name=".activities.DeletePlayersActivity"
             android:exported="false" />
         <activity
-            android:name=".delete_QRCodes"
+            android:name=".activities.DeleteCodesActivity"
             android:exported="false" />
         <activity
-            android:name=".OwnerMenu"
+            android:name=".activities.OwnerMenu"
             android:exported="false" />
         <activity
-            android:name=".ManageCodesActivity"
+            android:name=".activities.ManageCodesActivity"
             android:exported="false"
             android:label="@string/title_activity_manage_codes"
             android:theme="@style/Theme.QuiRky.NoActionBar" />
@@ -58,41 +58,41 @@
             android:value="YOUR_API_KEY" />
 
         <activity
-            android:name=".PlayerSearchActivity"
+            android:name=".activities.PlayerSearchActivity"
             android:exported="false" />
         <activity
-            android:name=".ViewQRActivity"
+            android:name=".activities.ViewQRActivity"
             android:exported="false"
-            android:parentActivityName=".StartingPageActivity" />
+            android:parentActivityName=".activities.StartingPageActivity" />
         <activity
-            android:name=".MapActivity"
+            android:name=".activities.MapActivity"
             android:exported="false"
             android:label="@string/title_activity_map"
             />
         <activity
-            android:name=".CodeScannerActivity"
+            android:name=".activities.CodeScannerActivity"
             android:exported="false"
-            android:parentActivityName=".StartingPageActivity" />
+            android:parentActivityName=".activities.StartingPageActivity" />
         <activity
-            android:name=".StartingPageActivity"
+            android:name=".activities.StartingPageActivity"
             android:exported="false" />
         <activity
-            android:name=".ViewCommentsActivity"
+            android:name=".activities.ViewCommentsActivity"
             android:exported="false" />
         <activity
-            android:name=".GenerateActivity"
+            android:name=".activities.GenerateActivity"
             android:exported="false"
-            android:parentActivityName=".StartingPageActivity" />
+            android:parentActivityName=".activities.StartingPageActivity" />
         <activity
-            android:name=".EditProfileActivity"
+            android:name=".activities.EditProfileActivity"
             android:exported="false"
-            android:parentActivityName=".StartingPageActivity" />
+            android:parentActivityName=".activities.StartingPageActivity" />
         <activity
-            android:name=".ProfileViewerActivity"
+            android:name=".activities.ProfileViewerActivity"
             android:exported="false"
-            android:parentActivityName=".StartingPageActivity" />
+            android:parentActivityName=".activities.StartingPageActivity" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
