+++ QuiRky/app/src/main/java/com/example/quirky/CommentList.java
@@ -4,15 +4,14 @@ import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.quirky.models.Comment;
+
 import java.util.ArrayList;
-import java.util.Collections;
 
 /**
  * Adapter class to map a Comment to a Recycler View.
