+++ QuiRky/app/src/main/java/com/example/quirky/InputUnameLoginFragment.java
@@ -3,7 +3,6 @@ package com.example.quirky;
 import android.app.AlertDialog;
 import android.app.Dialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -12,6 +11,8 @@ import android.widget.EditText;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
+
+import com.example.quirky.activities.CodeScannerActivity;
 
 /**
  * @author Jonathen Adsit
