+++ QuiRky/app/src/main/java/com/example/quirky/activities/GenerateActivity.java
@@ -1,4 +1,10 @@
-package com.example.quirky;
+/*
+ * Copyright (c) 2022. CMPUT301W22T29
+ * Subject to MIT License
+ * See full terms at https://github.com/CMPUT301W22T29/QuiRky/blob/main/LICENSE
+ */
+
+package com.example.quirky.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -8,6 +14,11 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.Toast;
+
+import com.example.quirky.controllers.MemoryController;
+import com.example.quirky.controllers.QRCodeController;
+import com.example.quirky.R;
+import com.example.quirky.controllers.DatabaseController;
 
 /**
  * Activity to generate QRCodes
