+++ QuiRky/app/src/main/java/com/example/quirky/activities/PlayerSearchActivity.java
@@ -4,7 +4,7 @@
  * See full terms at https://github.com/CMPUT301W22T29/QuiRky/blob/main/LICENSE
  */
 
-package com.example.quirky;
+package com.example.quirky.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -18,6 +18,14 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.ProgressBar;
+
+import com.example.quirky.ListeningList;
+import com.example.quirky.OnAddListener;
+import com.example.quirky.models.Profile;
+import com.example.quirky.QRAdapter;
+import com.example.quirky.R;
+import com.example.quirky.RecyclerClickerListener;
+import com.example.quirky.controllers.DatabaseController;
 
 import java.util.ArrayList;
