+++ QuiRky/app/src/main/java/com/example/quirky/activities/ProfileViewerActivity.java
@@ -1,13 +1,28 @@
-package com.example.quirky;
+/*
+ * Copyright (c) 2022. CMPUT301W22T29
+ * Subject to MIT License
+ * See full terms at https://github.com/CMPUT301W22T29/QuiRky/blob/main/LICENSE
+ */
+
+package com.example.quirky.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import com.example.quirky.ListeningList;
+import com.example.quirky.OnAddListener;
+import com.example.quirky.R;
+import com.example.quirky.activities.EditProfileActivity;
+import com.example.quirky.activities.ManageCodesActivity;
+import com.example.quirky.controllers.DatabaseController;
+import com.example.quirky.controllers.LeaderBoardController;
+import com.example.quirky.controllers.MemoryController;
+import com.example.quirky.models.Profile;
 
 /**
  * Activity to view a profile
