+++ QuiRky/app/src/main/java/com/example/quirky/controllers/DatabaseController.java
@@ -1,20 +1,22 @@
-package com.example.quirky;
+/*
+ * Copyright (c) 2022. CMPUT301W22T29
+ * Subject to MIT License
+ * See full terms at https://github.com/CMPUT301W22T29/QuiRky/blob/main/LICENSE
+ */
 
-import android.nfc.Tag;
+package com.example.quirky.controllers;
+
 import android.util.Log;
 
+import com.example.quirky.ListeningList;
+import com.example.quirky.models.Profile;
+import com.example.quirky.models.QRCode;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QuerySnapshot;
-import com.google.firebase.firestore.WriteBatch;
 
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 /**
  * @author Jonathen Adsit
