+++ QuiRky/app/src/main/java/com/example/quirky/controllers/MemoryController.java
@@ -1,6 +1,14 @@
-package com.example.quirky;
+/*
+ * Copyright (c) 2022. CMPUT301W22T29
+ * Subject to MIT License
+ * See full terms at https://github.com/CMPUT301W22T29/QuiRky/blob/main/LICENSE
+ */
+
+package com.example.quirky.controllers;
 
 import android.content.Context;
+
+import com.example.quirky.models.Profile;
 
 import java.io.File;
 import java.io.FileOutputStream;
