+++ QuiRky/app/src/main/java/com/example/quirky/models/Profile.java
@@ -1,6 +1,14 @@
-package com.example.quirky;
+/*
+ * Copyright (c) 2022. CMPUT301W22T29
+ * Subject to MIT License
+ * See full terms at https://github.com/CMPUT301W22T29/QuiRky/blob/main/LICENSE
+ */
+
+package com.example.quirky.models;
 
 import androidx.annotation.NonNull;
+
+import com.example.quirky.controllers.ProfileController;
 
 import java.io.Serializable;
 import java.util.ArrayList;
