+++ QuiRky/app/src/main/res/layout/activity_change_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditProfileActivity">
+    tools:context=".activities.EditProfileActivity">
 
     <TextView
         android:layout_width="wrap_content"
