+++ QuiRky/app/src/main/res/layout/activity_code_scanner.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CodeScannerActivity">
+    tools:context=".activities.CodeScannerActivity">
 
     <Button
         android:id="@+id/scan_button"
