+++ QuiRky/app/src/main/res/layout/activity_delete_qrcodes.xml
@@ -10,7 +10,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".delete_QRCodes"
+    tools:context=".activities.DeleteCodesActivity"
     android:orientation="vertical">
     <TextView
         android:id="@+id/textView"
