+++ QuiRky/app/src/main/res/layout/activity_generate.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GenerateActivity">
+    tools:context=".activities.GenerateActivity">
 
     <ImageView
         android:id="@+id/qrfield"
