+++ QuiRky/app/src/main/res/layout/activity_leader_board.xml
@@ -10,7 +10,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LeaderBoardActivity">
+    tools:context=".activities.LeaderBoardActivity">
 
     <TextView
         android:layout_width="wrap_content"
