+++ QuiRky/app/src/main/res/layout/activity_manage_codes.xml
@@ -10,7 +10,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ManageCodesActivity"
+    tools:context=".activities.ManageCodesActivity"
     android:orientation="vertical">
     <TextView
         android:id="@+id/manage_codes_title"
