+++ QuiRky/app/src/main/res/layout/activity_player_search.xml
@@ -10,7 +10,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PlayerSearchActivity">
+    tools:context=".activities.PlayerSearchActivity">
 
     <TextView
         android:id="@+id/search_activity_title"
