+++ QuiRky/app/src/main/res/layout/activity_qrcode_controller.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".QRCodeController">
+    tools:context=".controllers.QRCodeController">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
