+++ QuiRky/app/src/main/res/layout/activity_starting_page.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".StartingPageActivity"
+    tools:context=".activities.StartingPageActivity"
     android:layout_height="match_parent"
     android:layout_width="match_parent">
