+++ QuiRky/app/src/main/res/layout/activity_view_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfileViewerActivity">
+    tools:context=".activities.ProfileViewerActivity">
 
     <TextView
         android:id="@+id/profile_name"
